pls work

+++ apps/lend/src/components/reward-banner.tsx
@@ -59,7 +59,7 @@ export const Banner = () => {
     actionType: TransactionActionType.CLAIMING_REWARDS,
     onSuccess: () => refetch(),
   });
-  const formattedRewards = formatEther(rewards ?? 0n);
+  const formattedRewards = formatEther((rewards ?? 0n) as bigint);
   const { isSmall, numericValue: formattedBgt } =
     formatAmountSmall(formattedRewards);
